added instruction size, RIP register
Former-commit-id: dc06ea2e0c9e02c5495c024e4e43c25c54905c53

+++ libMEDSannotation/include/VirtualOffset.hpp
@@ -24,7 +24,7 @@
 #include <string>
 #include <sstream>
 
-typedef unsigned ApplicationAddress;
+typedef unsigned long long ApplicationAddress;
 
 #define DEFAULT_LIBRARY_NAME "a.out"
